requirements.txt
biopython==1.82
python-magic==0.4.27